Divide beam modules into several files
